Refactor code for detection and segment
