+++ .gitignore
@@ -1,3 +1,13 @@
+# Personal
+.idea
+*.engine
+*.pt
+*.pth
+*.onnx
+*.jpg
+*.png
+*.bmp
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
