+++ export-det.py
